refactor(pokeinfo): restore 4-value actuals and split test files
- StatActuals now holds all 4 values (maxPlus/max/min/minMinus)
  with null for HP's maxPlus/minMinus
- Embed displays all 4 values per stat with section header
- Split tests to match source structure:
  index.test.ts, view-model.test.ts, embed.test.ts
